refactor(ors_control_app): extract routes/diagnostics
Move /api/diagnostics/{logs,env,probe,logs/clear} into routes/diagnostics.ts
Express Router. server/index.ts mounts it via app.use(createDiagnosticsRouter(APP_VERSION)).

Phase 1 of holistic-control-app-refactor-final, route extraction step 2.
Lines down 3,160 -> 3,085. tsc passes cleanly.
